+++ R/add-functions.R
@@ -165,7 +165,13 @@ preproc_reference <- function(reference, cnames, n_rows) {
 #' ped_info(ped) %>% add_hazard(pam, type = "response")
 #' ped_info(ped) %>% add_cumu_hazard(pam)
 #' @export
-add_hazard <- function(
+add_hazard <- function(newdata, object, ...) {
+  UseMethod("add_hazard", object)
+}
+
+#' @rdname add_hazard
+#' @export
+add_hazard.default <- function(
   newdata,
   object,
   reference = NULL,
@@ -189,7 +195,7 @@ add_hazard <- function(
       newdata <- newdata %>% select(-one_of(rm.vars))
   }
 
-  get_hazard(newdata, object, reference = reference,
+  get_hazard(object, newdata, reference = reference,
     ci = ci, type = type, se_mult = se_mult, ci_type = ci_type,
     time_var = time_var, ...)
 
@@ -201,9 +207,14 @@ add_hazard <- function(
 #' @importFrom stats model.frame
 #' @importFrom mgcv predict.gam predict.bam
 #' @keywords internal
-get_hazard <- function(
-  newdata,
+get_hazard <- function(object, newdata, ...) {
+  UseMethod("get_hazard", object)
+}
+
+#' @rdname get_hazard
+get_hazard.default <- function(
   object,
+  newdata,
   reference = NULL,
   ci        = TRUE,
   type      = c("response", "link"),
@@ -241,7 +252,7 @@ get_hazard <- function(
     newdata <- newdata %>% mutate(hazard = exp(.data$hazard))
   }
 
-  newdata
+  newdata %>% arrange(.data[[time_var]], .by_group = TRUE)
 
 }
 
@@ -318,7 +329,7 @@ get_cumu_hazard <- function(
   if (ci) {
     if (ci_type == "default" | ci_type == "delta") {
       vars_exclude <- c(vars_exclude, "se", "ci_lower", "ci_upper")
-      newdata <- get_hazard(newdata, object, type = "response", ci = ci,
+      newdata <- get_hazard(object, newdata, type = "response", ci = ci,
         ci_type = ci_type, time_var = time_var, se_mult = se_mult, ...)
       if (ci_type == "default") {
         mutate_args <- mutate_args %>%
@@ -332,15 +343,15 @@ get_cumu_hazard <- function(
       }
     } else {
       if (ci_type == "sim") {
-        newdata <- get_hazard(newdata, object, type = "response", ci = FALSE,
+        newdata <- get_hazard(object, newdata, type = "response", ci = FALSE,
           time_var = time_var, ...)
         newdata <- split(newdata, group_indices(newdata)) %>%
           map_dfr(get_sim_ci_cumu, object = object, nsim = nsim, ...)
       }
     }
   } else {
-    newdata <- newdata %>%
-      get_hazard(object, type = "response", ci = ci,
+    newdata <-
+      get_hazard(object, newdata, type = "response", ci = ci,
         ci_type = ci_type, time_var = time_var, se_mult = se_mult, ...)
   }
   newdata <- newdata %>%
@@ -432,7 +443,7 @@ get_surv_prob <- function(
   if (ci) {
     if (ci_type == "default" | ci_type == "delta") {
       vars_exclude <- c(vars_exclude, "se", "ci_lower", "ci_upper")
-      newdata <- get_hazard(newdata, object, type = "response", ci = ci,
+      newdata <- get_hazard(object, newdata, type = "response", ci = ci,
         ci_type = ci_type, time_var = time_var,  se_mult = se_mult, ...)
       if (ci_type == "default") {
         mutate_args <- mutate_args %>%
@@ -446,15 +457,15 @@ get_surv_prob <- function(
       }
     } else {
       if (ci_type == "sim") {
-        newdata <- get_hazard(newdata, object, type = "response", ci = FALSE,
+        newdata <- get_hazard(object, newdata, type = "response", ci = FALSE,
           time_var = time_var, ...)
         newdata <- split(newdata, group_indices(newdata)) %>%
           map_dfr(get_sim_ci_surv, object = object, nsim = nsim, ...)
       }
     }
   } else {
-    newdata <- newdata %>%
-      get_hazard(object = object, type = "response", ci = FALSE,
+    newdata <-
+      get_hazard(object = object, newdata, type = "response", ci = FALSE,
         time_var = time_var, ...)
   }
 
@@ -562,6 +573,7 @@ add_delta_ci_surv <- function(newdata, object, se_mult = 2, ...) {
 }
 
 #' Calculate simulation based confidence intervals
+#'
 #' @keywords internal
 #' @importFrom mvtnorm rmvnorm
 #' @importFrom stats coef
@@ -610,6 +622,127 @@ get_sim_ci_surv <- function(newdata, object, alpha = 0.05, nsim = 100L, ...) {
 
   newdata$surv_lower <- apply(sim_fit_mat, 1, quantile, probs = alpha / 2)
   newdata$surv_upper <- apply(sim_fit_mat, 1, quantile, probs = 1 - alpha / 2)
+
+  newdata
+
+}
+
+
+## Cumulative Incidence Function (CIF) for competing risks data
+
+
+#' Add cumulative incidence function to data
+#'
+#' @inheritParams add_hazard
+#' @param alpha The alpha level for confidence/credible intervals.
+#' @param n_sim Number of simulations (draws from posterior of estimated coefficients)
+#' on which estimation of CIFs and their confidence/credible intervals will be
+#' based on.
+#' @param cause_var Character. Column name of the 'cause' variable.
+#'
+#' @export
+add_cif <- function(
+  newdata,
+  object,
+  ...) {
+
+  UseMethod("add_cif", object)
+
+}
+
+
+#' @rdname add_cif
+#' @export
+add_cif.default <- function(
+  newdata,
+  object,
+  ci        = TRUE,
+  overwrite = FALSE,
+  alpha     = 0.05,
+  n_sim     = 500L,
+  cause_var = "cause",
+  time_var = NULL,
+  ...) {
+
+  coefs        <- coef(object)
+  V            <- object$Vp
+  sim_coef_mat <- mvtnorm::rmvnorm(n_sim, mean = coefs, sigma = V)
+
+  map_dfr(
+    split(newdata, group_indices(newdata)),
+    ~get_cif(
+      newdata = .x, object = object, ci = ci, alpha = alpha, n_sim = n_sim,
+      cause_var = cause_var, coefs = coefs, V = V, sim_coef_mat = sim_coef_mat,
+      time_var = time_var, ...)
+  )
+
+}
+
+#' Calculate CIF for one cause
+#'
+#' @keywords internal
+get_cif <- function(newdata, object, ...) {
+
+  UseMethod("get_cif", object)
+
+}
+
+#' @rdname get_cif
+#' @keywords internal
+get_cif.default <- function(
+  newdata,
+  object,
+  ci,
+  time_var,
+  alpha,
+  n_sim,
+  cause_var,
+  coefs,
+  V,
+  sim_coef_mat,
+  ...) {
+
+  is_gam <- inherits(object, "gam")
+  if (is.null(time_var)) {
+    time_var <- ifelse(is_gam, "tend", "interval")
+  } else {
+    assert_string(time_var)
+    assert_choice(time_var, colnames(newdata))
+  }
+
+
+  causes_model <- as.factor(object$attr_ped$risks)
+  cause_data   <- unique(newdata[[cause_var]])
+
+  if(length(cause_data) > 1) {
+    stop("Did you forget to group by cause?")
+  }
+
+  hazards <- map(
+    causes_model,
+    ~ {
+        .df <- mutate(newdata, cause = .x) %>%
+          arrange(.data[[time_var]], .by_group = TRUE)
+        X <- predict(object, .df, type = "lpmatrix")
+        apply(sim_coef_mat, 1, function(z) exp(X %*% z))
+      }
+    )
+  overall_survivals <- apply(
+    Reduce("+", hazards),
+    2,
+    function(z) exp(-cumsum(z * newdata[["intlen"]])))
+  names(hazards) <- causes_model
+  # calculate cif
+  hazard           <- hazards[[cause_data]]
+  # Value of survival just prior to time-point
+  survival         <- overall_survivals - 1e-20
+  hps              <- hazard * survival
+  cifs             <- apply(hps, 2, function(z) cumsum(z * newdata[["intlen"]]))
+  newdata[["cif"]] <- rowMeans(cifs)
+ if(ci) {
+    newdata[["cif_lower"]] <- apply(cifs, 1, quantile, alpha/2)
+    newdata[["cif_upper"]] <- apply(cifs, 1, quantile, 1-alpha/2)
+  }
 
   newdata