+++ R/as-ped.R
@@ -4,6 +4,7 @@
 #' \code{pammtools} package. Two main applications must be distinguished:
 #' \enumerate{
 #'  \item Transformation of standard time-to-event data.
+#'  \item Transformation of left-truncated time-to-event data.
 #'  \item Transformation of time-to-event data with time-dependent covariates (TDC).
 #' }
 #' For the latter, the type of effect one wants to estimate is also
@@ -24,7 +25,9 @@
 #' @param formula A two sided formula with a \code{\link[survival]{Surv}} object
 #' on the left-hand-side and covariate specification on the right-hand-side (RHS).
 #' The RHS can be an extended formula, which specifies how TDCs should be transformed
-#' using specials \code{concurrent} and \code{cumulative}.
+#' using specials \code{concurrent} and \code{cumulative}. The left hand-side can
+#' be in start-stop-notation. This, however, is only used to create left-truncated
+#' data and does not support the full functionality.
 #' @param cut Split points, used to partition the follow up into intervals.
 #' If unspecified, all unique event times will be used.
 #' @param max_time If \code{cut} is unspecified, this will be the last
@@ -55,11 +58,21 @@ as_ped.data.frame <- function(
   cut          = NULL,
   max_time     = NULL,
   tdc_specials = c("concurrent", "cumulative"),
-  censor_code = 0L,
+  censor_code  = 0L,
+  transition   = character(),
+  timescale    = c("gap", "calendar"),
+  min_events   = 1L,
   ...) {
 
   status_error(data, formula)
   assert_subset(tdc_specials, c("concurrent", "cumulative"))
+
+  if (test_character(transition, min.chars = 1L, min.len = 1L)) {
+    ped <- as_ped_recurrent(data = data, formula = formula, cut = cut,
+      max_time = max_time, tdc_specials = tdc_specials, censor_code = censor_code,
+      transition = transition, timescale = timescale, min_events = min_events, ... )
+    return(ped)
+  }
 
   event_types <- get_event_types(data, formula, censor_code)
   if (length(event_types) > 1) {
@@ -167,8 +180,8 @@ as_ped.list <- function(
 
     }
   }
-  attr(ped, "time_var") <- get_lhs_vars(formula)[1]
-  attr(ped, "status_var") <- get_lhs_vars(formula)[2]
+  lhs_vars <- get_lhs_vars(formula)
+  attr(ped, "time_var") <- lhs_vars[1]
   attr(ped, "trafo_args")$formula <- formula
 
   ped
@@ -241,7 +254,7 @@ as_ped_cr <- function(
     }
   )
   if(length(cut) > 1 & combine) {
-    cut <- list(do.call(union, cut))
+    cut <- list(reduce(cut, union))
   }
 
   ped <- map2(
@@ -264,15 +277,15 @@ as_ped_cr <- function(
     ped <- do.call(rbind, ped)
     class(ped) <- c("ped_cr_union", "ped_cr", class(ped))
     attr(ped, "intvars") <- c(attr(ped, "intvars"), "cause")
-    attr(ped, "breaks") <- cut
+    attr(ped, "breaks") <- if (length(cut) ==1) unlist(cut) else cut
   } else {
     class(ped) <- c("ped_cr_list", "ped_cr", "ped", class(ped))
-    names(ped) <- event_types
+    names(ped) <- paste0("cause = ", event_types)
     attributes(ped)$trafo_args$id <- attributes(ped[[1]])$trafo_args$id
     attributes(ped)$trafo_args$formula <- formula
   }
 
-  attr(ped, "trafo_args")[["cut"]] <- cut
+  attr(ped, "trafo_args")[["cut"]] <- if (length(cut) ==1) unlist(cut) else cut
   attr(ped, "trafo_args")[["combine"]] <- combine
   attr(ped, "trafo_args")[["censor_code"]] <- censor_code
   attr(ped, "risks") <- event_types
@@ -293,7 +306,66 @@ as_ped_cr <- function(
 get_event_types <- function(data, formula, censor_code) {
 
   lhs_vars <- get_lhs_vars(formula)
-  status_values <- unique(data[[lhs_vars[2]]]) %>% sort()
+  status_values <- unique(data[[lhs_vars[length(lhs_vars)]]]) %>% sort()
   status_values[status_values != censor_code]
+
+}
+
+
+
+#' Recurrent events trafo
+#'
+#' @examples
+#' \dontrun{
+#' data("cgd", package = "frailtyHL")
+#' cgd2 <- cgd %>%
+#'  select(id, tstart, tstop, enum, status, age) %>%
+#'  filter(enum %in% c(1:2))
+#' ped_re <- as_ped_recurrent(
+#'   formula = Surv(tstart, tstop, status) ~ age + enum,
+#'   data = cgd2,
+#'  transition = "enum")
+#' }
+#' @rdname as_ped
+#' @export
+#' @keywords internal
+as_ped_recurrent <- function(
+  data,
+  formula,
+  cut          = NULL,
+  max_time     = NULL,
+  tdc_specials = c("concurrent", "cumulative"),
+  censor_code  = 0L,
+  transition  = character(),
+  timescale    = c("gap", "calendar"),
+  min_events   = 1L,
+  ...
+) {
+
+  assert_character(transition, min.chars = 1L, min.len = 1L, any.missing = FALSE,
+    len = 1L)
+  assert_integer(min_events, lower = 1L, len = 1L)
+
+  status_error(data, formula)
+  assert_subset(tdc_specials, c("concurrent", "cumulative"))
+
+  rhs_vars <- get_rhs_vars(formula)
+  if (!(transition %in% rhs_vars)) {
+    formula <- add_to_rhs(formula, transition)
+  }
+
+  dots            <- list(...)
+  dots$data       <- data
+  dots$formula    <- get_ped_form(formula, data = data, tdc_specials = tdc_specials)
+  dots$cut        <- cut
+  dots$max_time   <- max_time
+  dots$transition <- transition
+  dots$min_events <- min_events
+  dots$timescale  <- timescale
+
+  ped <- do.call(split_data_recurrent, dots)
+  attr(ped, "time_var")   <- get_lhs_vars(dots$formula)[1]
+
+  return(ped)
 
 }