+++ R/tidiers.R
@@ -55,6 +55,10 @@ tidy_fixed.gam <- function(x, intercept=FALSE, ...) {
 #' @rdname tidy_fixed
 #' @importFrom tibble as_tibble
 #' @keywords internal
+#' @examples
+#' library(survival)
+#' gc <- coxph(Surv(days, status)~age + sex, data = tumor)
+#' tidy_fixed(gc)
 #' @export
 tidy_fixed.coxph <- function(x, ...) {
 
@@ -172,8 +176,8 @@ tidy_re <- function(x, keep=c("fit", "main", "xlab", "ylab"), ...) {
     re.df  <- do.call(cbind.data.frame, c(z, stringsAsFactors = FALSE))
     re.df$x <- qnorm(ppoints(length(re.df$fit))[order(order(re.df$fit))])
     # code to calculate qqslope and qqintercept from ?stats::qqline
-    yl <- quantile(re.df$fit, probs = c(0.1, 0.9), type = 7, names = FALSE)
-    xl <- qnorm(c(0.1, 0.9))
+    yl <- quantile(re.df$fit, probs = c(0.25, 0.75), type = 7, names = FALSE)
+    xl <- qnorm(c(0.25, 0.75))
     re.df$qqslope <- diff(yl) / diff(xl)
     re.df$qqintercept <- yl[1L] - re.df$qqslope * xl[1L]